added JWT token, role based authentication and Session Management

+++ src/main/java/com/RESTfullCRUD/BasicCRUD/service/impl/ProductServiceImpl.java
@@ -1,33 +1,44 @@
 package com.RESTfullCRUD.BasicCRUD.service.impl;
 
 import com.RESTfullCRUD.BasicCRUD.entity.Product;
+import com.RESTfullCRUD.BasicCRUD.entity.User;
+import com.RESTfullCRUD.BasicCRUD.exceptions.CustomException;
 import com.RESTfullCRUD.BasicCRUD.repository.ProductRepository;
 //import com.RESTfullCRUD.BasicCRUD.repository.VendorRepository;
+import com.RESTfullCRUD.BasicCRUD.dto.responseDTO.ProductAndVendorDTO;
+import com.RESTfullCRUD.BasicCRUD.repository.UserRepository;
 import com.RESTfullCRUD.BasicCRUD.service.ProductService;
+import com.RESTfullCRUD.BasicCRUD.utils.JwtTokenUtil;
 import lombok.extern.log4j.Log4j2;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.ObjectUtils;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
+import java.util.Optional;
 
 @Log4j2
 @Service
 public class ProductServiceImpl implements ProductService {
     @Autowired
     private ProductRepository productRepository;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private JwtTokenUtil jwtTokenUtil;
 
     @Override
-    public Product addProduct(Product product) throws Exception {
+    public Product addProduct(Product product, String token) throws Exception {
+        String vendorName = jwtTokenUtil.getUsernameFromToken(token.substring(7));
+        product.setVendor(userRepository.findMatchedUser(vendorName));
         return productRepository.save(product);
     }
 
     @Override
     public List<Product> fetchProducts() throws Exception {
         try {
-            List<Product> productList = new ArrayList<>();
+            List<Product> productList;
             productList =  productRepository.findAll();
             if (!productList.isEmpty()) {
                 return productList;
@@ -48,9 +59,14 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public void removeProductById(Long productId) {
-        productRepository.deleteById(productId);
-//        return null;
+    public void removeProductById(Long productId, String token) throws CustomException {
+        String vendorName = jwtTokenUtil.getUsernameFromToken(token.substring(7));
+        String vendorNameFromDB = productRepository.getVendorName(productId);
+        if((vendorNameFromDB.equals(vendorName))){
+            productRepository.deleteById(productId);
+        }else {
+            throw new CustomException("User not Authorized to remove product");
+        }
     }
 
     @Override
@@ -59,37 +75,35 @@ public class ProductServiceImpl implements ProductService {
     }
 
     @Override
-    public Product updateProduct(Long productId, Product product) {
+    public Product updateProduct(Long productId, Product product, String token) throws CustomException {
         Product oldProd = productRepository.findById(productId).get();
-
-        if(Objects.nonNull(product.getProdName()) && !"".equalsIgnoreCase(product.getProdName())){
-            oldProd.setProdName(product.getProdName());
+        String vendorName = jwtTokenUtil.getUsernameFromToken(token.substring(7));
+        if(oldProd.getVendor().getUserName().equals(vendorName)){
+            if(Objects.nonNull(product.getProdName()) && !"".equalsIgnoreCase(product.getProdName())){
+                oldProd.setProdName(product.getProdName());
+            }
+            if(Objects.nonNull(product.getProdType()) && !"".equalsIgnoreCase(product.getProdType())){
+                oldProd.setProdType(product.getProdType());
+            }
+            if(Objects.nonNull(product.getPrice()) && product.getPrice() != 0.0){
+                oldProd.setPrice(product.getPrice());
+            }
+            if(Objects.nonNull(product.getVendor()) && !ObjectUtils.isEmpty(product.getVendor())){
+                throw new CustomException("Can't Change vendor Details. Try Again!!");
+            }
+            return productRepository.save(oldProd);
         }
-        if(Objects.nonNull(product.getProdType()) && !"".equalsIgnoreCase(product.getProdType())){
-            oldProd.setProdType(product.getProdType());
-        }
-//        System.out.println(product.getPrice());
-        if(Objects.nonNull(product.getPrice()) && product.getPrice() != 0.0){
-            oldProd.setPrice(product.getPrice());
-        }
-        if(Objects.nonNull(product.getVendor()) && !ObjectUtils.isEmpty(product.getVendor())){
-//            if(Objects.nonNull(product.getVendor().getVendorName()) && !"".equalsIgnoreCase(product.getVendor().getVendorName())){
-//            }
-                oldProd.setVendor(product.getVendor());
-        }
-
-        return productRepository.save(oldProd);
+        throw new CustomException("This Vendor is not Authorized to change/update this product");
     }
 
     @Override
-    public List<Object> fetchProductsWithConditions(String vendorName, String city) {
-        return productRepository.findWithVendorAndCity(vendorName,city);
+    public List<ProductAndVendorDTO> fetchProductsWithConditions(String userName, String city) {
+        System.out.println(productRepository.findWithVendorAndCity(userName,city));
+        return productRepository.findWithVendorAndCity(userName,city);
     }
 
 //    @Override
-//    public Object fetchProductsWithCondition(String vendorName, String city) {
-//        return productRepository.findWithVendorAndCity(vendorName,city);
+//    public List<Object> fetchProductsWithConditions(FetchProductConditionsRequest fetchProductConditionsRequest) {
+//        return productRepository.findWithVendorAndCity(fetchProductConditionsRequest.getVendorName(),fetchProductConditionsRequest.getCity());
 //    }
-
-
 }
